embedded-chatmock/requirements.txt
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
flask==3.1.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
requests==2.32.5
PyYAML==6.0.2
urllib3==2.5.0
websockets==15.0.1
werkzeug==3.1.3
